use `long` for dict size

+++ src/getpy.cpp
@@ -199,7 +199,7 @@ struct Dict {
     }
 
 
-    int __len__ () {
+    long __len__ () {
         return __dict.size();
     }
